requirements.txt
asgiref==3.3.1
Django==2.2
django-tenant-schemas==1.10.0
ordered-set==4.0.2
psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
